consumer/requirements.txt
kafka-python
hdfs[avro]
